refactor: added useItem hook to abstract out fetching item by id
