+++ public/react/components/EditForm.jsx
@@ -1,10 +1,9 @@
 // Form to edit an item
 // Item props are
 import apiURL from "../api";
-
+import useItem from "../useItem";
 // redirect sends back to the id view after form is done
 import { useParams, useNavigate } from "react-router-dom";
-const itemPropList = ["title", "description", "price", "category", "image"];
 
 // Form should have option selector which sets state to whatever property needs to change
 import React, { useEffect, useState } from "react";
@@ -13,13 +12,7 @@ import { FormComponent } from "./FormComponent";
 export const EditForm = () => {
   const { id } = useParams();
   const redirect = useNavigate();
-  const [formObject, setFormObject] = useState(null);
-
-  useEffect(async () => {
-    const res = await fetch(`${apiURL}/items/${id}`);
-    const data = await res.json();
-    setFormObject(() => data);
-  }, [id]);
+  const item = useItem(id);
 
   const submitFunction = async (formObject) => {
     try {
@@ -37,8 +30,8 @@ export const EditForm = () => {
     }
   };
 
-  return formObject ? (
-    <FormComponent initialState={formObject} submitFunction={submitFunction} />
+  return item ? (
+    <FormComponent initialState={item} submitFunction={submitFunction} />
   ) : (
     <p>Loading...</p>
   );
